Create foundation for handling multiple errors
For better convenience, the long term goal is to try and parse as much
as possible and then report multiple errors at once. For this we will
have to replace faulty intermediate parse results with default
assumptions and collect the errors in a separate structure.

The complete switch from create_error() to append_error() and removing
of the node_parse_error from the ast_node collection will need multiple
steps though and should not be done in one single go. We also need to
extend the parsing tests accordingly and make sure, we can actually
handle multiple errors.