Implement basic global variables
We want to progress on libraries and printing, so we need some very
basic global-variable handling to deal with the heap_ptr. In this
implementation we ignore any variable type other than i32 (for example
no string support) and we currently assume, the initialiser is a single
value.
In theory and also with WASM a global could be initialised with an
expression and we should address that eventually.